refactor(storybook): split stories
